tests/wpt/tests/css/css-borders/tentative/parsing/box-shadow-position-invalid.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Borders and Box Decorations 4 Test: Parsing 'box-shadow-position' with invalid values</title>
<link rel="author" title="Sebastian Zartner" href="mailto:sebastianzartner@gmail.com">
<link rel="help" href="https://drafts.csswg.org/css-borders-4/#box-shadow-position">
<meta name="assert" content="This test checks that 'box-shadow-position' supports only valid keywords.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>

<script>
test_invalid_value("box-shadow-position", "auto");
test_invalid_value("box-shadow-position", "1 2");
test_invalid_value("box-shadow-position", "1px");
test_invalid_value("box-shadow-position", "1px 2px");
test_invalid_value("box-shadow-position", "1%");
</script>
